% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_xy.R
\name{classify_xy}
\alias{classify_xy}
\title{Detect spatial information in data columns}
\usage{
classify_xy(list, x_or_y)
}
\arguments{
\item{list}{A metajam list output for a single data entity}

\item{x_or_y}{"x" or "y", dimension to look for}
}
\value{
A logical vector
}
\description{
Detect spatial information in data columns
}
