% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleverName_shiny.R
\name{CleverName_shiny}
\alias{CleverName_shiny}
\title{Creating a graphical user interface for creating ggplot-graphs.}
\usage{
CleverName_shiny(dataset = NA)
}
\arguments{
\item{dataset}{A dataset (optional).}
}
\value{
A GUI for visualizing data from \code{dataset}.
}
\description{
To run use \code{CleverName_shiny()}
}
