% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_data_package.R
\name{copy_data_package}
\alias{copy_data_package}
\title{Copy data package from temporary directory to user specified directory}
\usage{
copy_data_package(copy.dir)
}
\arguments{
\item{copy.dir}{(character) Directory to which the data package will be copied.}
}
\value{
(data package) Data package contents as parsed by \code{metajam}.
}
\description{
This function copies the data package from the temporary directory
specified by paste0(tempdir(), '/data_package') and copies it to the user
specified directory.
}
\examples{
\dontrun{
# Download data package to temporary directory
output <- read_data_archived(data.pkg.doi = 'doi:10.18739/A2DP3X')

# Copy data package to user specified directory
copy_data_package(copy.dir = '/Users/csmith/Desktop/data_packages')
}


}
