% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_copy.R
\name{data_package_copy}
\alias{data_package_copy}
\title{Copy data package from temporary directory to user specified directory}
\usage{
data_package_copy(
    copy.dir
  )
}
\arguments{
\item{copy.dir}{(character) Directory to which the data package will be copied.}
}
\value{
(data package) Data package contents as parsed by \code{metajam}.
}
\description{
This function copies the data package from the temporary directory
specified by paste0(tempdir(), '/data_package') to a user
specified directory.
}
\examples{
\dontrun{
# Download data package to temporary directory
data_package_download('doi:10.18739/A2DP3X')

# Copy data package to user specified directory
copy_data_package('/Users/csmith/Desktop/data_packages')

# Clean up
data_package_remove()
}

}
