% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_read.R
\name{data_package_read}
\alias{data_package_read}
\title{Read a data package into R}
\usage{
data_package_read(
    data.pkg.path = NULL
  )
}
\arguments{
\item{data.pkg.path}{(character) Directory where a data package stored. Default (NULL) is
the temporary data package directory defined by
\code{paste0(tempdir(), 'data_package')}.}
}
\value{
(list) List of tibbles containing data and metadata. Each list object is
  named after the data object file name.
}
\description{
This function reads a data package downloaded with the function
\code{data_package_read()} to the R environment as a list object.
}
\details{
This function is a wrapper to \code{read_d1_files()} function of the 
  \href{https://github.com/NCEAS/metajam}{metajam} package but outputs an
  object with a slightly different format.
}
\examples{
\dontrun{
# Read data package from temporary directory -------------------------------

# Download data package to temporary directory
data_package_download(data.pkg.doi = 'doi:10.18739/A2DP3X')

# Read data package into R
pkg <- data_package_read()

# View data package contents
View(pkg)

# Clean up
data_package_remove()

# Read data package from local directory -----------------------------------

# Download data package to local directory
data_package_download(
  data.pkg.doi = 'doi:10.18739/A2DP3X',
  download.dir = '/Desktop/data_packages'
)

# Read data package into R
pkg <- data_package_read(
  data.pkg.path = '/Desktop/data_packages/doi_10.18739_A2DP3X'
)

# View data package contents
View(pkg)

# Clean up
data_package_remove('/Desktop/data_packages/doi_10.18739_A2DP3X')

}

}
