% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_remove.R
\name{data_package_remove}
\alias{data_package_remove}
\title{Remove a data package directory}
\usage{
data_package_remove(
    pkg.dir = NULL
  )
}
\arguments{
\item{pkg.dir}{(character) The directory to be removed. Default is the temporary 
directory specified by \code{paste0(tempdir(), '/data_package')}.}
}
\value{
(message) Message indicating whether or not the directory has been 
  removed.
}
\description{
This function removes a data package directory.
}
\examples{
\dontrun{
  # Remove the temporary data package directory
  data_package_remove()
}

}
