% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_wrapper.R
\name{data_package_wrapper}
\alias{data_package_wrapper}
\title{Download and read wrapper function}
\usage{
data_package_wrapper(data.pkg.doi, download.dir = NULL)
}
\arguments{
\item{data.pkg.doi}{(character) Directory where a data package stored. Default (NULL) is
the temporary data package directory defined by
\code{paste0(tempdir(), 'data_package')}.}

\item{download.dir}{(character) Directory to which the data package will be downloaded. 
Default is the temporary directory specified by 
\code{paste0(tempdir(), '/data_package')}.}
}
\value{
Returns a list of tibbles containing data and metadata for a given data
package. Attributes include information about the doi and download folder.
}
\description{
A wrapper for the \code{data_package_download} and \code{data_package_read}
functions. In some context, combining these functions makes them easier to
handle.
}
\details{
This function does two things:
1. It combines the \code{data_package_download()} and
\code{data_package_read()} functions into a single function.
2. It assigns attributes to the resulting data package, making it easier to
track where the data came from (doi) and where it is stored locally (folder).
}
\examples{
\dontrun{
#Download data to a temporary directory and then load it to the global
# environment.
example_package <- data_package_wrapper(data.pkg.doi = 'doi:10.18739/A2DP3X')

#Explicitly download to a temporary directory.
example_package <- data_package_wrapper(data.pkg.doi = 'doi:10.18739/A2DP3X',
  download.dir = tempdir())
  
#Download and load to current working directory
example_package <- data_package_wrapper(data.pkg.doi = 'doi:10.18739/A2DP3X',
  download.dir = getwd())

}


}
