% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapie_shiny.R
\name{datapie_shiny}
\alias{datapie_shiny}
\title{Creating a graphical user interface for creating ggplot-graphs.}
\usage{
datapie_shiny(dataset = NA)
}
\arguments{
\item{dataset}{A dataset (optional).}
}
\value{
A GUI for visualizing data from \code{dataset}.
}
\description{
To run use \code{datapie_shiny()}
}
