% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_package.R
\name{download_data_package}
\alias{download_data_package}
\title{Download data package to user specified directory}
\usage{
download_data_package(download.dir)
}
\arguments{
\item{download.dir}{(character) Directory to which the data package will be downloaded. 
Default is \code{paste0(tempdir(), '/data_package')}}
}
\description{
Download data package to user specified directory
}
\examples{

}
