% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqclocks_forDates.R
\name{freqclocks_forDates}
\alias{freqclocks_forDates}
\title{freqclocks_forDates}
\usage{
freqclocks_forDates(df, var, varname)
}
\arguments{
\item{df}{Define this parameter.}

\item{var}{Define this parameter.}

\item{varname}{Define this parameter.}
}
\value{
(ggplot clocks) Of the frequency of data occurring at different time frequencies, if Date type data.
}
\description{
This function creates ggplot clocks of the frequency of data occurring at different time frequencies, if Date type data.
}
