% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqclocks_forPOSIX.R
\name{freqclocks_forPOSIX}
\alias{freqclocks_forPOSIX}
\title{freqclocks_forPOSIX}
\usage{
freqclocks_forPOSIX(df, var, varname)
}
\arguments{
\item{data.df3}{Define this parameter.}
}
\value{
(ggplot clocks) Of the frequency of data occurring at different time frequencies, if POSIX type data.
}
\description{
This function creates ggplot clocks of the frequency of data occurring at different time frequencies, if POSIX type data.
}
