% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_entity_summary.R
\name{HackathonFunction1}
\alias{HackathonFunction1}
\title{Collection of functions to generate data-entity level summary information
By Jocelyn (An is documenting)}
\usage{
HackathonFunction1(data.df1)
}
\arguments{
\item{data.df1}{A dataframe containing the entity's data}
}
\value{

}
\description{
Collection of functions to generate data-entity level summary information
By Jocelyn (An is documenting)
}
