% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cat_variable_summary_df.R
\name{make_cat_variable_summary_df}
\alias{make_cat_variable_summary_df}
\title{make_cat_variable_summary_df}
\usage{
make_cat_variable_summary_df(var)
}
\arguments{
\item{var}{The categorical vector to be summarized}
}
\value{
A dataframe containing summary information.
}
\description{
Creates a dataframe that contains summary information (type, number of levels, num_valid, num_missing)
about categorical variables.
}
