% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_categorical_histogram.R
\name{make_categorical_histogram}
\alias{make_categorical_histogram}
\title{make_categorical_histogram}
\usage{
make_categorical_histogram(df, var, varname)
}
\arguments{
\item{df}{Define this parameter.}

\item{var}{Define this parameter.}

\item{varname}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
