% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_entity_report.R
\name{make_entity_report}
\alias{make_entity_report}
\title{A wrapper function to generate data entity-level summary plus variable-level reports for one entities}
\usage{
make_entity_report(data_entity_list)
}
\arguments{
\item{data_entity_list}{A metajam list output for a single data entity.}
}
\description{
A wrapper function to generate data entity-level summary plus variable-level reports for one entities
}
