% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_missing_plot.R
\name{make_missing_plot}
\alias{make_missing_plot}
\title{make_missing_plot}
\usage{
make_missing_plot(var)
}
\arguments{
\item{var}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
