% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_numeric_histogram.R
\name{make_numeric_histogram}
\alias{make_numeric_histogram}
\title{make_numeric_histogram}
\usage{
make_numeric_histogram(df, var, varname)
}
\arguments{
\item{df}{A dataframe}

\item{var}{A numeric column contained in df that will be graphed}

\item{varname}{A string containing the name of the column to be graphed. 
This is used only for labeling the graph, so alternative 
wording can be used, if desired.}
}
\value{
A ggplot2 object containing the histogram and density plot for that column
}
\description{
Creates a histogram with density plot for numeric vectors
}
