% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_numeric_variable_summary_df.R
\name{make_numeric_variable_summary_df}
\alias{make_numeric_variable_summary_df}
\title{make_numeric_variable_summary_df}
\usage{
make_numeric_variable_summary_df(var)
}
\arguments{
\item{var}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
