% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_variable_summary_df.R
\name{make_time_variable_summary_df}
\alias{make_time_variable_summary_df}
\title{make_time_variable_summary_df}
\usage{
make_time_variable_summary_df(var)
}
\arguments{
\item{var}{Define this parameter.}
}
\value{
(data frame) A summary table of time (POSIX, Date) objects.
}
\description{
This function writes a summary table of time (POSIX, Date) objects
}
