% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_variable_report.R
\name{make_vector_of_numeric_variable_names}
\alias{make_vector_of_numeric_variable_names}
\title{Collection of functions to create variable-specific report}
\usage{
make_vector_of_numeric_variable_names(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
By John (An is documenting)
}
