% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_missing_code.R
\name{match_names}
\alias{match_names}
\title{Match data and metadata attribute names.
This function matches indices of column names in the data and attributeName's specified in metadata.}
\usage{
match_names(
    entity_list
    )
}
\arguments{
\item{entity_list}{(list) A list object containing information on a single data entity in metajam output format.}
}
\value{
(vector) A numeric vector with row indices in attribute metadata to match column indices in data.
}
\description{
Match data and metadata attribute names.
This function matches indices of column names in the data and attributeName's specified in metadata.
}
