% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_wrapper.R
\name{rda_wrapper}
\alias{rda_wrapper}
\title{A wrapper for read_data_archive}
\usage{
rda_wrapper(data.pkg.doi, download.dir = NULL)
}
\arguments{
\item{data.pkg.doi}{The doi of the package being downloaded.}

\item{download.dir}{The download directory from the read_data_archived
function.}
}
\value{

}
\description{
A wrapper for read_data_archive
}
