% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_archived.R
\name{read_data_archived}
\alias{read_data_archived}
\title{Read an archived data package into R as a list object (THIS FUNCTION IS DEPRECATED. PLEASE USE data_package_download() and data_package_read() INSTEAD.)}
\usage{
read_data_archived(data.pkg.doi, download.dir = NULL)
}
\arguments{
\item{data.pkg.doi}{(character) DOI of a data package available in the DataONE network
(e.g. "doi:10.18739/A2DP3X").}

\item{download.dir}{(character) Directory to which the data package will be downloaded. 
Default is \code{paste0(tempdir(), '/data_package')}}
}
\value{
(data package) Data package objects (as rendered by \code{metajam}) in the
  download directory.
  (list of data objects) List of data objects created by \code{metajam}. 
  Readable data objects are defined by in the \code{metajam} package.
  Tabular data is read into tibbles.
}
\description{
This function downloads and reads a data package available in the 
\href{https://search.dataone.org/data}{DataONE} network to the R environment 
as a list object.
}
\details{
This function is a wrapper to the \code{download_d1_data_pkg()} and 
  \code{read_d1_data_pkg()} functions of the 
  \href{https://github.com/NCEAS/metajam}{metajam} package, but configured
  to output a list object usable by this project.
}
\examples{
\dontrun{
# Download data package to temporary directory and read
output <- read_data_archived(data.pkg.doi = 'doi:10.18739/A2DP3X')

# View data package content
View(output)
}

}
