% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data_package_directory.R
\name{remove_data_package_directory}
\alias{remove_data_package_directory}
\title{Remove the temporary data package directory}
\usage{
remove_data_package_directory(download.dir = paste0(tempdir(),
  "/data_package"))
}
\arguments{
\item{download.dir}{(character) The directory to be removed.}
}
\value{
(message) Message indicating whether or not the directory has been 
  removed.
}
\description{
This function removes the temporary data package directory created
by \code{download_data_package()} and to which the data package contents are
downloaded to. This task is required before a new data package is downloaded 
by the user.
}
\details{
The temporary data package directory is located at 
  \code{paste0(tempdir(), '/data_package')}
}
\examples{
\dontrun{
  # Remove default data package directory
  remove_data_package_directory()
}

}
