% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_detective.R
\name{space_detective}
\alias{space_detective}
\title{the space detective}
\usage{
space_detective(list)
}
\arguments{
\item{list}{Metajam output for a single data entity}
}
\value{
A character vector whose two elements are named "lat_col" and "lon_col", indicating the names of columns containing lat and lon information respectively, or an error message if not detected or if we aren't equipped to deal with the number of hits we got.
}
\description{
Wrapper for \code{\link{classify_xy}}. Runs \code{\link{classify_xy}}, counts number of hits in each dimension, then decides what to depending on those two numbers.
}
