% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_plot.R
\name{space_plot}
\alias{space_plot}
\title{the Space plot}
\usage{
space_plot(space_cols = space_cols, df, var)
}
\arguments{
\item{space_cols}{A character vector whose two elements are named "lat_col" and "lon_col". Outpur from \code{\link{space_detective}}}

\item{var}{Name of variable to make data availability and summary. Remember to enclose in quotes.}
}
\value{
For all variables, function returns a data availability plot over the study area. For categorical column types, function returns an extra plot showing most prevalent level of that variable in each spatial grid cell over the study area. For numeric column types, function returns an extra plot showing mean of that variable in each spatial grid cell over the study area. Lat and lon are binned; there are 50 bins normalized to attempt to enforce equal scales along each dimension.
}
\description{
Create variable specific spatial heatmaps to show data availability and data summary over the geographical study area.
}
