% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_report_complete.R
\name{static_report_complete}
\alias{static_report_complete}
\title{Wrapper function to generate complete HTML report: entity-level  plus variable-level reports for one entity.}
\usage{
static_report_complete(entity_list, output_path)
}
\arguments{
\item{entity_list}{(list) A list object containing information on a single data entity in metajam output format.}

\item{output_path}{(character) Path to save complete HTML report to. If "ask", a Rstudio pop-up window appears allowing choice of save directory (requires RStudio >= 1.1.287).}
}
\value{
A HTML file with complete static on the chosen data entity.
}
\description{
Wrapper function to generate complete HTML report: entity-level  plus variable-level reports for one entity.
}
