% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_report_entity.R
\name{static_report_entity}
\alias{static_report_entity}
\title{Generate entity-level report.}
\usage{
static_report_entity(entity_df)
}
\arguments{
\item{entity_df}{(data.frane) (data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}
}
\value{
A list containing entity-level report items.
}
\description{
Generate entity-level report.
}
