% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_report_variable.R
\name{static_report_variable}
\alias{static_report_variable}
\title{Generate variable-level plots.}
\usage{
static_report_variable(entity_df, varname, space_cols = space_cols)
}
\arguments{
\item{entity_df}{(data.frame) A data.frame containing entity-level data. The "data" child element of an entity-level list object in metajam output format.}

\item{varname}{(character) Name of variable of interest.}

\item{space_cols}{(character vector) Output from \code{space_detective}. Where data has explicit spatial columns, a two-element character vector indicating column names containing longitudes and latitudes. Otherwise, spatial plots will not be generated.}
}
\value{
A list object containing variable name and relevant plots.
}
\description{
Generate variable-level plots.
}
