% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable1.R
\name{SummaryTable1}
\alias{SummaryTable1}
\title{SummaryTable1}
\usage{
SummaryTable1(data.df1)
}
\arguments{
\item{data.df1}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
