% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable2.R
\name{SummaryTable2}
\alias{SummaryTable2}
\title{SummaryTable2}
\usage{
SummaryTable2(data.df3)
}
\arguments{
\item{data.df3}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
