% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTable3.R
\name{SummaryTable3}
\alias{SummaryTable3}
\title{SummaryTable3}
\usage{
SummaryTable1(data.df2)
}
\arguments{
\item{data.df2}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Table to display categorical summary information....
}
