% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_detector.R
\name{time_detector}
\alias{time_detector}
\title{time.detector}
\usage{
time_detector(df)
}
\arguments{
\item{df}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
This function takes character-class columns and attempts to coerce them into Date-class, if reasonable test whether a column conforms to standard date formats.
}
