% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_missing_code.R
\name{use_missing_code}
\alias{use_missing_code}
\title{Substitudes missing codes with NAs.
This function performs column-wise substitution of NAs for missing values codes, if specified in attribute metadata.}
\usage{
use_missing_code(
    entity_list
    )
}
\arguments{
\item{entity_list}{(list) A list object containing information on a single data entity in metajam output format.}
}
\value{
(list) If attribute metadata is present and missing codes are specified: a list object but with NAs substituted for missing codes. Otherwise, a list object identical to input.
}
\description{
Substitudes missing codes with NAs.
This function performs column-wise substitution of NAs for missing values codes, if specified in attribute metadata.
}
