% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_package_shiny_handler.R
\name{data_package_shiny_handler}
\alias{data_package_shiny_handler}
\title{A data_package_* handler for the shiny app}
\usage{
data_package_shiny_handler(data.pkg.doi, current.data = NULL,
  download.dir = NULL)
}
\arguments{
\item{data.pkg.doi}{The doi of the package being downloaded.}

\item{current.data}{Data current}

\item{download.dir}{The download directory from the read_data_archived
function.}
}
\value{
Returns a list of tibbles containing data and metadata for a given data
package. Attributes include information about the doi and download folder.
}
\description{
A data_package_* handler for the shiny app
}
