% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_categorical_histogram.R
\name{make_categorical_histogram}
\alias{make_categorical_histogram}
\title{make_categorical_histogram}
\usage{
make_categorical_histogram(df, var, varname)
}
\arguments{
\item{df}{A dataframe}

\item{var}{The col.name drawn from df of the vector to be graphed}

\item{varname}{A string containing the name of the column to be graphed. 
This is used only for labeling the graph, so alternative 
wording can be used if desired.}
}
\value{
A ggplot2 object containing the histogram for that column
}
\description{
Creates a histogram of the frequency with which different codes occur for categorical variables.
}
