% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_missing_plot.R
\name{make_missing_plot}
\alias{make_missing_plot}
\title{make_missing_plot}
\usage{
make_missing_plot(var)
}
\arguments{
\item{var}{Name of the vector whose missing values are to be graphed}
}
\value{
A ggplot2 graph object
}
\description{
Creates a plot that shows where missing values occur in the data. 
The plot appears as a bar with vertical lines representing valid or NA observations for a single variable.
}
