% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_numeric_variable_summary_df.R
\name{make_numeric_variable_summary_df}
\alias{make_numeric_variable_summary_df}
\title{make_numeric_variable_summary_df}
\usage{
make_numeric_variable_summary_df(var)
}
\arguments{
\item{var}{Name of the vector containing the numbers to be summarized}
}
\value{
A dataframe containing the summary information.
}
\description{
Creates a dataframe containing summary (median,max,min,num_valid,num_missing) information for numerical variables.
}
