% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vector_of_numeric_variable_names.R
\name{make_vector_of_numeric_variable_names}
\alias{make_vector_of_numeric_variable_names}
\title{make_vector_of_numeric_variable_names}
\usage{
make_vector_of_numeric_variable_names(df)
}
\arguments{
\item{df}{Define this parameter.}
}
\value{
What does this function return?
}
\description{
Describe what this function does
}
